+++ resources/js/Pages/Frontend/Profile/Profile.vue
@@ -1,8 +1,4 @@
 <template>
-  <div>
-    <!-- <Dashboard /> -->
-  </div>
-
   <v-container>
     <v-row no-gutters>
       <v-col cols="auto" md="12">
@@ -83,109 +79,12 @@
                        
                         <v-expansion-panel-text>
                           <v-sheet class="ma-2 pa-6">
-                            <v-card elevation="0">
-
-                              <v-card-text>
-                                <v-row>
-                                  <v-col cols="auto" md="12">
-                                    <label>Address:</label>
-                                    <v-text-field v-model="editedItem.address" required hide-details
-                                      :label="customerAddress.endereco">
-                                    </v-text-field>
-
-                                  </v-col>
-
-                                 
-                                  <v-col cols="auto" sm="5">
-                                    <label>Complement:</label>
-                                    <v-text-field v-model="editedItem.complemento" required hide-details
-                                      :label="customerAddress.complemento"
-                                      :placeholder="customerAddress.complemento"></v-text-field>
-                                  </v-col>
-                                  
-                                  <v-col cols="auto" md="3">
-                                    <label>N°</label>
-                                    <v-text-field v-model="editedItem.number" required hide-details :label="customerAddress.number"
-                                      :placeholder="customerAddress.number"></v-text-field>
-                                  </v-col>
-
-                                  <v-col cols="auto" sm="4">
-                                    <label>Bairro:</label>
-                                    <v-text-field v-model="editedItem.bairro" required hide-details :label="customerAddress.bairro"
-                                      :placeholder="customerAddress.bairro"></v-text-field>
-                                  </v-col>
-                                </v-row>
-                                <v-row>
-
-                                  <v-col cols="auto" md="3">
-                                    <label>UF:</label>
-                                    <v-select v-model="editedItem.uf" :label="customerAddress.uf" :items="ufs" item-title="uf"
-                                      placeholder="UF" return-object single-line :placeholder="customerAddress.uf">
-
-                                    </v-select>
-
-                                  </v-col>
-                                  <v-col cols="auto" sm="4">
-                                    <label>State:</label>
-                                    <v-text-field v-model="editedItem.estado" required hide-details :placeholder="uf.state"
-                                      :label="customerAddress.estado"></v-text-field>
-                                  </v-col>
-                                </v-row>
-                                <v-row>
-                                  <v-col cols="auto" sm="4">
-                                    <label>Postal code:</label>
-                                    <v-text-field v-model="editedItem.zip_code" v-maska:[options] :label="customerAddress.cep"
-                                      :placeholder="customerAddress.cep"></v-text-field>
-
-
-                                  </v-col>
-                                  <v-col cols="auto" sm="4">
-                                    <label>Cidade:</label>
-                                    <v-text-field v-model="editedItem.cidade" required hide-details :label="customerAddress.cidade"
-                                      :placeholder="customerAddress.cidade"></v-text-field>
-                                  </v-col>
-
-                                </v-row>
-
-                              </v-card-text>
-                              <v-row>
-                                <v-col cols="auto" sm="4">
-                                  <label>Pais:</label>
-                                  <v-select v-model="editedItem.country" aria-required hide-details :label="customerAddress.pais"
-                                    :items="countryItems" :placeholder="customerAddress.pais">
-
-                                  </v-select>
-                                </v-col>
-                                <v-col cols="auto" sm="5">
-                                  <label>Phone:</label>
-                                  <v-text-field v-model="editedItem.phone" required hide-details :label="customerAddress.telefone"
-                                    v-maska:[phoneOptions] :placeholder="customerAddress.telefone"></v-text-field>
-                                </v-col>
-                                <v-col col="8" sm="6">
-                                  <v-btn-group>
-                                    <v-btn :disabled="loadingUpdate" :loading="loadingUpdate" class="text-none mb-4"
-                                    color="indigo-darken-3" size="large" variant="plain"
-                                    @click="editAddress(customerAddress)" v-if="Object.keys(customerAddress).length >= 1">
-                                    Update
-                                  </v-btn>
-                                  <v-btn v-if="Object.keys(customerAddress).length >= 1" class="text-none mb-4" color="error" size="large" variant="plain" 
-                                    @click="openDialogRemoveAddress(customerAddress)">
-                                    Remove address
-                                  </v-btn>
-                                  <v-btn :disabled="loading" :loading="loading" block class="text-none mb-4"
-                                    color="indigo-darken-3" size="x-large" variant="flat" @click="loading = !loading"
-                                    v-else="object.keys(customerAddress).length == 0">
-                                    Save and continue
-                                  </v-btn>
-                                
-                                  </v-btn-group>
-                                  
-                                </v-col>
-                              </v-row>
-
-
-                            </v-card>
-                            <div>
+                            <FirstAddress 
+                              :editedItem="this.editedItem"
+                              :customerAddress="this.customerAddress"
+                              :uf="this.uf"/>
+                            
+                              <div>
                               <ProfileUpload
                                 v-model="dialogImage"
                                 v-if="dialogImage"
@@ -261,6 +160,7 @@ const phoneMask = ref('');
 <script>
 
 import axios from "axios";
+import FirstAddress from './partials/FirstAddress.vue';
 import ProfileUpload from '../Dialogs/ProfileImage.vue'
 import DeleteImageProfile from "../Profile/profileImage/deleteImageProfile.vue";
 import ProfileCustomerDialog from '../Profile/partials/ProfileCustomerDialog.vue';
@@ -269,6 +169,7 @@ import RemoveAddressDialog from '../Profile/partials/removeAddress.vue';
 export default {
   emits: ['close-dialog', 'delete-image'],
   components: {
+    FirstAddress,
     ProfileUpload,
     ProfileCustomerDialog,
     DeleteImageProfile,
@@ -303,18 +204,7 @@ export default {
     dialogRemoveAddress: false,
     imageId: -1,
     imageRemove: false,
-    ufs: [{
-      state: 'Distrito Federal', uf: 'DF',
-      state: 'Goiás', uf: 'Go',
-      state: 'Maranhão', uf: 'MA',
-    }
-
-      //     'RJ',
-      //    'SP',
-    ],
-    countryItems: [
-      'Brazil'
-    ],
+   
     editedItem: {
       address: '',
       number: '',
@@ -329,21 +219,6 @@ export default {
     },
 
   }),
-  watch: {
-    loading(val) {
-      if (!val) return;
-      this.save();
-      setTimeout(() => (this.loading = false, 2000));
-    },
-    loadingUpdate(val) {
-      if (!val) return;
-      this.update();
-      setTimeout(() => (this.loadingUpdate = false, 20000));
-    },
-    dialogImage(val) {
-      val || this.closeImageDialog();
-    }
-  },
   methods: {
     getCustomer() {
       axios.get('/customer')
@@ -411,37 +286,37 @@ export default {
     closeImageDialog() {
       this.dialogImage = false;
     },
-    openDialogRemoveAddress(item){
-     // this.editedItem = Object.assign({}, item);
-    //  this.editedIndex = this.customerAddress.indexOf(item);
-      this.dialogRemoveAddress = true;
-    },
+    // openDialogRemoveAddress(item){
+    //  // this.editedItem = Object.assign({}, item);
+    // //  this.editedIndex = this.customerAddress.indexOf(item);
+    //   this.dialogRemoveAddress = true;
+    // },
     uploadProfileImage() {
       this.dialogImage = true;
     },
-    save() {
-      const data = {
-        address: this.editedItem.address,
-        number: this.editedItem.number,
-        complemento: this.editedItem.complemento,
-        bairro: this.editedItem.bairro,
-        uf: this.editedItem.uf.uf,
-        state: this.editedItem.uf.state,
-        postal_code: this.editedItem.zip_code,
-        city: this.editedItem.cidade,
-        country: this.editedItem.country,
-        phone: this.editedItem.phone,
-      };
-      axios.post('/address/save', data)
-        .then((response) => {
-          
-          return this.customerAddress = response.data;
-        })
-        .catch((response) => {
-          return alert('Error :' + response);
-        });
+      // save() {
+      //   const data = {
+      //     address: this.editedItem.address,
+      //     number: this.editedItem.number,
+      //     complemento: this.editedItem.complemento,
+      //     bairro: this.editedItem.bairro,
+      //     uf: this.editedItem.uf.uf,
+      //     state: this.editedItem.uf.state,
+      //     postal_code: this.editedItem.zip_code,
+      //     city: this.editedItem.cidade,
+      //     country: this.editedItem.country,
+      //     phone: this.editedItem.phone,
+      //   };
+      //   axios.post('/address/save', data)
+      //     .then((response) => {
+            
+      //       return this.customerAddress = response.data;
+      //     })
+      //     .catch((response) => {
+      //       return alert('Error :' + response);
+      //     });
 
-    },
+      // },
     editAddress() {
       return this.update();
     },
